Explicitly ensure a nil `so-long-action' does nothing
This ensures that our hard-coded "Do nothing" option is accurate, even
if the user was to define an action with a key of nil.

Also provide a default :value for the key to a new custom action, so
that it does not default to nil.  (We could :validate against it, but
that seems like overkill.)