Change `so-long-change-major-mode' to :before advice for `so-long-mode'
This way we only run the code when we are calling `so-long-mode'.

Previously this was run in `change-major-mode-hook', which meant that
it was redundant during most calls (as most major mode changes are not
changing to `so-long-mode').

This means we do not need to test this-command, so we can more simply
check `so-long--calling' in this function to decide how to behave.
(This then has a flow-on effect in `so-long-mode' itself, which also
tests `so-long--calling'.)

When Emacs bug 35351 is resolved we will use that new feature in Emacs
versions 27+.