Compatibility fix for Emacs versions < 26.1
Prevent `so-long-mode-revert' from triggering a file-local 'mode'.

In 26.1 `run-mode-hooks' calls (hack-local-variables 'no-mode), using
the new HANDLE-MODE argument behaviour.  In prior versions of Emacs we
call `hack-local-variables' in `so-long-mode-revert' in order to
achieve the same result, and we need to mimic the `no-mode'
functionality to ensure that after restoring `so-long-original-mode'
we do not then clobber it with a file-local mode.